Add reworked FixedMessage to all filters
Merge all changes from ChatModerator/fixed-fix

Update all filters to the new FixedMessage format

Add initial spec for the FixedMessage object

Fix NPE on matcher instantiation in several filters

`fixed` parameter seems to be broken?

Standardize Pattern variable and method name

Update to work with the new FixedMessages

Fixed matching on original message in all filters

Closes OvercastNetwork/ChatModerator#7

Update spec for a FixedMessage

Update all filters to new FixedMessage spec

Oh god whitespace

Standardize the Instant for all violations on a message

Add timeSent parameter to the spec for FixedMessage

Change getViolationsFor() --> getViolationsForType() in the
PlayerViolationManager

Add getViolationsForTime() in the PlayerViolationManager

Update all Filters to create a Violation with the same Instant as when the
message was sent in chat.

Rework how messages are sent according to FixedMessage

Add toString to the FixedMessage class